Rename package lib -> commands
